extract bench setup/teardown
